Use props for all shapes, cleanup

+++ include/PaceLib/graphics/gui/Props.h
@@ -3,6 +3,7 @@
 #include "../utils/SDL2/SDL_FontCache.h"
 #include <SDL2/SDL.h>
 #include <string>
+#include <vector>
 
 namespace PaceLib {
 
@@ -50,20 +51,20 @@ struct CircleProp {
 };
 
 struct TriangleProp {
-  float x1;
-  float y1;
-  float x2;
-  float y2;
-  float x3;
-  float y3;
+  int x1;
+  int y1;
+  int x2;
+  int y2;
+  int x3;
+  int y3;
   SDL_Color color;
 };
 
 struct LineProp {
-  float x1;
-  float y1;
-  float x2;
-  float y2;
+  int x1;
+  int y1;
+  int x2;
+  int y2;
   SDL_Color color;
 };
